requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
azure-core==1.38.0
pyyaml==6.0.1
azure-eventhub==5.15.1
azure-storage-blob==12.28.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
frozenlist==1.8.0
idna==3.11
isodate==0.7.2
multidict==6.7.0
numpy==2.4.1
pandas==2.3.3
propcache==0.4.1
pyarrow==22.0.0
pycparser==2.23
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8.1
tenacity==9.1.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
yarl==1.22.0
